PR #83 w/ modifications
