+++ src/locale/locales/es/common.json
@@ -34,7 +34,7 @@
     "send": "enviar",
     "receive": "recibir",
     "receive_info": "Neuron elige una nueva dirección de recepción para mejorar la privacidad. Por favor, diríjase a la Libreta de Direcciones si desea utilizar una dirección de recepción previamente utilizada.",
-    "copy": "Copia",
+    "copy": "Copiar",
     "import_wallet": "Importar wallet",
     "enter_mnemonic": "Introducir mnemonic",
     "mnemonic_input_text": "Escribe una palabra y presiona espacio o enter para añadirla o toca una para eliminarla.",
