+++ src/module/transaction/component/core/ReceiveModal/ReceiveModal.tsx
@@ -2,17 +2,12 @@ import { Col, createBackdrop, ExposedBackdropProps, Typography, useToast } from 
 import CardNavigatorModal from "module/common/component/navigation/CardNavigatorModal/CardNavigatorModal";
 import { useTranslate } from "module/common/hook/useTranslate";
 import QRCode from "module/transaction/component/display/QRCode/QRCode";
-import styled from "@peersyst/react-native-styled";
 import useSelectedWallet from "module/wallet/hook/useSelectedWallet";
 import { serviceInstancesMap } from "module/wallet/state/WalletState";
 import useSelectedNetwork from "module/settings/hook/useSelectedNetwork";
 import Button from "module/common/component/input/Button/Button";
 import * as Clipboard from "expo-clipboard";
 import Container from "module/common/component/display/Container/Container";
-
-const TextAddress = styled(Typography, { textTransform: "uppercase" })(() => ({
-    width: "100%",
-}));
 
 const ReceiveModal = createBackdrop<ExposedBackdropProps>(({ close, ...rest }) => {
     const t = useTranslate();
@@ -25,10 +20,6 @@ const ReceiveModal = createBackdrop<ExposedBackdropProps>(({ close, ...rest }) =
     const copyToClipboard = () => {
         Clipboard.setString(address || "");
         showToast(t("address_copied"), { type: "success" });
-        closeModal();
-    };
-
-    const closeModal = () => {
         close();
     };
 
@@ -36,13 +27,13 @@ const ReceiveModal = createBackdrop<ExposedBackdropProps>(({ close, ...rest }) =
         <CardNavigatorModal navbar={{ back: true, title: t("receive") }} {...rest}>
             <Col gap={"8%"} flex={1} justifyContent="center">
                 <QRCode />
-                <Typography textAlign="center" variant="body2">
+                <Typography textAlign="center" variant="body3Regular">
                     {t("receive_info")}
                 </Typography>
                 <Container>
-                    <TextAddress variant="body1" textAlign="center">
+                    <Typography variant="body2Strong" textAlign="center" style={{ textTransform: "uppercase" }}>
                         {address}
-                    </TextAddress>
+                    </Typography>
                 </Container>
                 <Button variant="primary" fullWidth onPress={() => copyToClipboard()}>
                     {t("copy")}
